refactor: split PianoRoll (933→<600 lines) into sub-components
